+++ wipesrc/wipeout98/draw.cpp
@@ -24,7 +24,7 @@
 #include "trackasm.h"
 #include "effects.h"
 #include "fileops.h"
-#include "menus.h"  
+#include "menus.h"
 #include "hiscores.h"
 #include "hud.h"
 #include "fri.h"
@@ -37,24 +37,24 @@ extern TrackCamera		ctrlcamera;
 
 
 char TrackDirection
-( 
-	TrackCamera*		camera 
+(
+	TrackCamera*		camera
 	)
 {
 	SVECTOR        reference;
-	long           angle;
+	int32_t           angle;
 	TrackSection   *nextSection;
-	
-	
+
+
 	nextSection = camera->section->nextSection.ptr;
-	
-	reference.vx = (short)(nextSection->centre.vx - camera->section->centre.vx );
-	reference.vy = (short)(nextSection->centre.vy - camera->section->centre.vy );
-	reference.vz = (short)(nextSection->centre.vz - camera->section->centre.vz );
-	
+
+	reference.vx = (int16_t)(nextSection->centre.vx - camera->section->centre.vx );
+	reference.vy = (int16_t)(nextSection->centre.vy - camera->section->centre.vy );
+	reference.vz = (int16_t)(nextSection->centre.vz - camera->section->centre.vz );
+
 	angle = -ratan2(reference.vx, reference.vz) + camera->hdg ;
 	angle = ang(angle);
-	
+
 	if((angle >= -512) && (angle < 512))
 	{
 		return (ShipForwards);
@@ -77,67 +77,67 @@ char TrackDirection
 
 
 
-long  count = 0;
+int32_t  count = 0;
 
 
-long  c = 0;
+int32_t  c = 0;
 
 void DrawLandscape
-( 
-	ShipData			*playerShip, 
+(
+	ShipData			*playerShip,
 	TrackCamera		*camera
 	)
 {
-	short          i;
+	int16_t          i;
 	VECTOR         result;
-	long           flag;
-	TrackSection* 	section; 
-	long				extent;
+	int32_t           flag;
+	TrackSection* 	section;
+	int32_t				extent;
 	SVECTOR        centre;
-	
+
 	for ( i=0; i<MaxUnique; i++ )
 	{
 		sharedPrimCount[ i ] = 0;
 	}
-	
+
 	section = camera->section;
-	
+
 	if ( section->objectCount > 0 )
 	{
 		for ( i=0; i<section->objectCount; i++ )
 		{
 			objectTable[ section->objects[ i ] ]->skeleton->update = 1;
-			
+
 			SetSkeletonTransform( objectTable[ section->objects[ i ] ]->skeleton );
-			
+
 			extent = objectTable[ section->objects[ i ] ]->extent;
-			
+
 			centre.vx = 0;
 			centre.vy = 0;
 			centre.vz = 0;
-			
+
 			RotTrans( &centre, &result, &flag );
-			
+
 			result.vz += extent;
-			
+
 			if ( result.vz == 0 ) result.vz = 1;
-			
+
 			result.vx = ( result.vx * 160 ) / result.vz;
 			result.vy = ( result.vy * 160 ) / result.vz;
-			
+
 			extent = ( extent * 160 ) / result.vz;
-			
-			if 
+
+			if
 				(
-				( result.vx >= -( extent + 160 ) ) && 
+				( result.vx >= -( extent + 160 ) ) &&
 				( result.vx <=  ( extent + 160 ) ) &&
 				( result.vy >= -( extent + 120 ) ) &&
 				( result.vy <=  ( extent + 120 ) )
 				)
 			{
-				
+
 #if 1
-				
+
 				if ( playerShip->attr & VIEW_IN )
 				{
 					DrawObjectAsm( objectTable[ section->objects[ i ] ], ctrlcamera.camPos, playerShip->roll, 0 );
@@ -146,9 +146,9 @@ void DrawLandscape
 				{
 					DrawObjectAsm( objectTable[ section->objects[ i ] ], ctrlcamera.camPos, 0, 0 );
 				}
-				
+
 #else
-				
+
 				if ( playerShip->attr & VIEW_IN )
 				{
 					TransformLibraryObject( objectTable[ section->objects[ i ] ], &glo.camPos, playerShip->roll );
@@ -157,12 +157,12 @@ void DrawLandscape
 				{
 					TransformLibraryObject( objectTable[ section->objects[ i ] ], &glo.camPos, 0 );
 				}
-				
-#endif          
-			}      
-		}         
-	}            
-	
-	count += 1;  
-	
+
+#endif
+			}
+		}
+	}
+
+	count += 1;
+
 }
